requirements.txt
fastapi
uvicorn
python-multipart
pillow
aiofiles
tqdm
matplotlib
seaborn
pandas
opencv-python
requests
torch
torchvision
ffmpeg